[#19 #28] Integrate add_track_id_to_playlist and search_track_name_on_spotify
Add tracks to database at user_playlist_tracks
Add track_artist to Track model
Functions add_track_id_to_playlist and search_track_name_on_spotify intended to work with user_playlist_tracks and route to existing urls
Handle search results url removed

Issues:
 - Authentication
 - Some functions can be moved to spotify/utils.py file (example: append track results at user_playlist_tracks function and spotify/utils.py append_track_data perform similar functions and may not need authentication)
 - Requires run and test upon authentication (i.e. no problems show in console but without running cannot be sure)
 - Messages may require formatting
