+++ spotify/views.py
@@ -1,4 +1,3 @@
-from typing import NoReturn
 from django.shortcuts import render, redirect
 from django.contrib.auth.decorators import login_required
 from django.shortcuts import get_object_or_404
@@ -8,8 +7,9 @@ from django.conf import settings
 import spotipy
 import os
 
-from .models import Playlist
+from .models import Playlist, Track
 from room.models import Room
+from .utils import append_track_data
 
 
 # Default parameter value to start while loop
@@ -24,8 +24,8 @@ def session_cache_path(request):
 def authorize_with_spotify(request, spotify_code=None):
     cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path(request))
     auth_manager = spotipy.oauth2.SpotifyOAuth(scope='playlist-read-private',
-                                                cache_handler=cache_handler,
-                                                show_dialog=True)
+                                               cache_handler=cache_handler,
+                                               show_dialog=True)
 
     # If we have a code from Spotify then grab the autho token and refresh token and bung in .spotify_caches directory.
     if spotify_code:        
@@ -47,204 +47,6 @@ def authorize_with_spotify(request, spotify_code=None):
 
     return render(request, 'user_playlists.html', context)
 
-# TO DO: Refactor functions
-@login_required
-def get_playlist_items(request, auth_manager, PLAYLIST_ID):
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
-
-    MAX_ITEMS_LIMIT    = 100
-    OFFSET             = 0
-    OFFSET_UPPER_LIMIT = 10
-    playlist_items     = None
-
-    while ( ( NEXT != 'null' | NEXT != None ) & ( OFFSET <= OFFSET_UPPER_LIMIT ) ):
-        try:
-            returned_data = spotify.playlist_items(playlist_id=PLAYLIST_ID, 
-                                                   fields=None, 
-                                                   limit=MAX_ITEMS_LIMIT, 
-                                                   offset=OFFSET, 
-                                                   market='GB', 
-                                                   additional_types='track')
-            LIMIT = returned_data['limit']
-            NEXT  = returned_data['next']
-            TOTAL = returned_data['total']
-
-            if TOTAL > 0:
-                playlist_items = append_track_data(returned_data, LIMIT)
-            else:
-                break
-        except:        
-            messages.error(request, 'ERROR ' + request['error']['status'] + ': ' + 'Playlist Items ' + request['error']['message'])
-        else:
-            messages.success(request, 'SUCCESS: Playlist Items Fetched')
-        
-        OFFSET += 1
-
-        if OFFSET > OFFSET_UPPER_LIMIT:
-            messages.error(request, 'ERROR: There are over 1000 songs in your playlist? Let''s stop here!')
-            break
-        else:
-            pass
-    
-    return playlist_items
-
-# Scenario 1: Add a new track using a valid Spotify track name - Happy Path
-# Scenario 2a: Search for a new track using an ambiguous Spotify track name - Happy Path
-# Scenario 2b: Search for a new track to add to the playlist - Happy Path
-@login_required
-def search_track_name_on_spotify(request, auth_manager, QUERY, PLAYLIST_ID):
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
-
-    MAX_SEARCH_LIMIT   = 100
-    OFFSET             = 0
-    OFFSET_UPPER_LIMIT = 1
-    search_results     = None
-
-    while ( ( NEXT != 'null' | NEXT != None ) & ( OFFSET <= OFFSET_UPPER_LIMIT ) ):
-        try:
-            returned_data = spotify.search(q=QUERY, limit=MAX_SEARCH_LIMIT, offset=OFFSET, type='track', market='GB')
-
-            LIMIT = returned_data['limit']
-            NEXT  = returned_data['next']
-            TOTAL = returned_data['total']
-            
-            if TOTAL > 0:
-                search_results = append_track_data(request, returned_data, LIMIT)
-            else:
-                messages.info(request, 'INFO: No tracks found.')
-                break
-        except:
-            messages.error(request, 'ERROR: Unknown error fetching track names.')
-
-        OFFSET += 1
-
-        if OFFSET > OFFSET_UPPER_LIMIT:
-            messages.error(request, 'ERROR: There are over 100 search results! Let''s stop here!')
-            break
-        else:
-            pass
-
-    return handle_search_results(search_results, PLAYLIST_ID)
-
-@login_required
-def handle_search_results(request, search_results, PLAYLIST_ID):
-    if len(search_results['track_id']) == 1:
-        # Scenario 1: Add a new track using a valid Spotify track name - Happy Path
-        TRACK_ID = search_results['track_id']
-        add_track_id_to_playlist(request, PLAYLIST_ID, TRACK_ID) 
-    elif len(search_results['track_id']) > 1:
-        # Scenario 2a: Search for a new track using an ambiguous Spotify track name - Happy Path
-        # Scenario 2b: Search for a new track to add to the playlist - Happy Path
-        context = search_results
-        return render(request, 'handle_search_results.html', context)  
-    else:
-        messages.info(request, 'INFO: No track available.')
-        return redirect('room.html')  
-    # Scenario 3: Add a new track using an invalid Spotify track name - Unhappy Path
-    # Handled by add_track_id_to_playlist
-    # Scenario 4: Add a duplicate track using a Spotify ID - Unhappy Path
-    # Handled by add_track_id_to_playlist
-
-@login_required
-def add_track_id_to_playlist(request, auth_manager, PLAYLIST_ID, TRACK_ID):
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
-
-    # TO DO: Do we need to do this again? 
-    # Should pass parameter to check how we arrived here, to check when playlist_track_ids were last fetched
-    playlist_items = get_playlist_items(PLAYLIST_ID)
-    playlist_track_ids = playlist_items['track_id']
-    new_playlist_track_ids = playlist_track_ids
-
-    if TRACK_ID not in playlist_track_ids:
-        try: 
-            # Scenario 1: Add a new track using a valid Spotify ID - Happy Path
-            # {
-            #         "snapshot_id": "MzgsNjM1NDNkN2I0MGNlNzBjNDc0NGYxODAyOWVhZmEyNjFlZWQxYTZiZg=="
-            # }
-            snapshot_id = spotify.playlist_add_items(playlist_id=PLAYLIST_ID, items=TRACK_ID, position=None)
-        except:
-            # Scenario 2: Add a new track using an invalid Spotify ID - Unhappy Path
-            # {
-            #     "error": {
-            #     "status": 400,
-            #     "message": "Payload contains a non-existing ID"
-            # }
-            messages.error(request, 'ERROR ' + request['error']['status'] + ': ' + request['error']['message'])
-        else:
-            new_playlist_track_ids = get_playlist_items(PLAYLIST_ID)
-            messages.success(request, 'SUCCESS? Add track id to playlist attempted.')
-    else:
-        # Scenario 3: Add a duplicate track using a Spotify ID - Unhappy Path
-        messages.error(request, 'ERROR: Track already on playlist')
-        return redirect('room.html') 
-
-    context = {
-        'PLAYLIST_ID': PLAYLIST_ID,
-        'playlist_track_ids': new_playlist_track_ids
-    }
-
-    return render(request, 'room.html', context)
-
-@login_required
-def append_track_data(request, returned_data, LIMIT):
-    # TRACK_ARTIST, TRACK_NAME and TRACK_ALBUM are intended for view
-    # TRACK_ID is used for adding to a playlist        
-    track_artist = []
-    track_name   = []
-    track_id     = []
-    track_album  = []
-
-    index = 0        
-    while index < LIMIT:
-        try:
-            track_artist.append(returned_data['items'][index]['track']['artists']['name'])
-            track_name.append(returned_data['items'][index]['track']['name']) 
-            track_id.append(returned_data['items'][index]['track']['id'])
-            track_album.append(returned_data['items'][index]['track']['album']['name'])
-            index += 1  
-        except IndexError:
-            break
-        except:
-            messages.error(request, 'ERROR: Unknown Error compiling track ids, artist names and track names.')
-        else:
-            context = {
-                    'track_artist': track_artist,
-                    'track_name':   track_name,
-                    'track_id':     track_id,
-                    'track_album':  track_album
-            }
-    
-    return context
-    
-@login_required
-def user_playlist_tracks(request, playlist_id, playlist_name):
-    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path(request))
-    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
-
-    spotify = spotipy.Spotify(auth_manager=auth_manager)
-    user_playlist_tracks = spotify.playlist_items(playlist_id,
-                                                    offset=0,
-                                                    fields='items.track.id, items.track.name')
-    
-    list_of_track_ids = []
-
-    for track in user_playlist_tracks['items']:
-        list_of_track_ids.append(track['track']['id'])
-
-    results = spotify.tracks(list_of_track_ids)
-    
-    track_name_artist = []
-    for track in results['tracks']:
-        track_name_artist.append(track['name'] + ' - ' + track['artists'][0]['name'])              
-
-    context = {
-        'playlist_id': playlist_id,
-        'playlist_name': playlist_name,
-        'track_name_artist': track_name_artist
-    }    
-    
-    return render(request, 'user_playlist_tracks.html', context)
-
 @login_required
 def add_playlist_to_room(request, playlist_id, playlist_name):
     room = get_object_or_404(Room, pk=request.user.active_room_id, created_by=request.user, status=Room.ACTIVE, members__in=[request.user])
@@ -255,3 +57,170 @@ def add_playlist_to_room(request, playlist_id, playlist_name):
     messages.info(request, "Your playlist was added to your room.")
 
     return render(request, 'grooveboard.html')
+
+@login_required
+def user_playlist_tracks(request, playlist_id, playlist_name):
+    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path(request))
+    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
+    spotify = spotipy.Spotify(auth_manager=auth_manager)
+
+    user_playlist_tracks = spotify.playlist_items(playlist_id,
+                                                  offset=0,
+                                                  fields='items.track.id, items.track.name')
+    
+    list_of_track_ids = []
+
+    for track in user_playlist_tracks['items']:
+        list_of_track_ids.append(track['track']['id'])
+
+    results = spotify.tracks(list_of_track_ids)
+    
+    track_name_artist = []
+
+    # TO DO: clear Tracks model rebuild user playlist tracks everytime?
+    for track in results['tracks']:
+        track_name_artist.append(track['name'] + ' - ' + track['artists'][0]['name'])
+        
+        # TO DO: place in .utils file
+        # for all results (tracks), if not already in database, add to database; else, pass
+        if Track.objects.filter(track_id=track['id']).count() == 0: 
+            Track(playlist=playlist_name, track_id=track['id'], track_name=track['name'], track_artist = track['artists'][0]['name']).save()
+        else:
+            pass
+
+    context = {
+        'playlist_id': playlist_id,
+        'playlist_name': playlist_name,
+        'track_name_artist': track_name_artist
+    }    
+    
+    return render(request, 'user_playlist_tracks.html', context)
+
+# TO DO: Re-factor with class functions
+# @login_required
+# def get_playlist_items(request, playlist_id):
+#     spotify = spotipy.Spotify(auth_manager=auth_manager)
+#     MAX_ITEMS_LIMIT    = 100
+#     OFFSET             = 0
+#     OFFSET_UPPER_LIMIT = 10
+#     while ( ( NEXT != 'null' | NEXT != None ) & ( OFFSET <= OFFSET_UPPER_LIMIT ) ):
+#         try:
+#             returned_data = spotify.playlist_items(playlist_id, 
+#                                                    fields=None, 
+#                                                    limit=MAX_ITEMS_LIMIT, 
+#                                                    offset=OFFSET, 
+#                                                    market='GB', 
+#                                                    additional_types='track')
+#             LIMIT = returned_data['limit']
+#             NEXT  = returned_data['next']
+#             TOTAL = returned_data['total']
+#             if TOTAL > 0:
+#                 playlist_items = append_track_data(returned_data, LIMIT)
+#             else:
+#                 break
+#         except:        
+#             messages.error(request, 'ERROR ' + request['error']['status'] + ': ' + 'Playlist Items ' + request['error']['message'])
+#         else:
+#             messages.success(request, 'SUCCESS: Playlist Items Fetched')       
+#         OFFSET += 1
+#         if OFFSET > OFFSET_UPPER_LIMIT:
+#             messages.error(request, 'ERROR: There are over 1000 songs in your playlist? Let''s stop here!')
+#             break
+#         else:
+#             pass
+
+# Scenario 1: Add a new track using a valid Spotify track name - Happy Path
+# Scenario 2a: Search for a new track using an ambiguous Spotify track name - Happy Path
+# Scenario 2b: Search for a new track to add to the playlist - Happy Path
+@login_required
+def search_track_name_on_spotify(request, auth_manager, query, playlist_id, playlist_name):
+    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path(request))
+    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
+    spotify = spotipy.Spotify(auth_manager=auth_manager)
+
+    MAX_SEARCH_LIMIT   = 100
+    OFFSET             = 0
+    OFFSET_UPPER_LIMIT = 1
+
+    search_results     = None
+
+    while ( ( NEXT != 'null' | NEXT != None ) & ( OFFSET <= OFFSET_UPPER_LIMIT ) ):
+        try:
+            returned_data = spotify.search(q=query, limit=MAX_SEARCH_LIMIT, offset=OFFSET, type='track', market='GB')
+
+            LIMIT = returned_data['limit']
+            NEXT  = returned_data['next']
+            TOTAL = returned_data['total']
+            
+            if TOTAL > 0:
+                search_results = append_track_data(request, returned_data, LIMIT)
+            else:
+                messages.info(request, 'INFO: No tracks found.')
+                return redirect('room.html')
+        except:
+            messages.error(request, 'ERROR: Unknown error fetching track names.')
+            return redirect('room.html')
+
+        OFFSET += 1
+
+        if OFFSET > OFFSET_UPPER_LIMIT:
+            messages.info(request, 'INFO: There are over 100 search results! Let''s stop here!')
+            return redirect('room.html')
+        else:
+            pass
+
+    add_search_result_to_room(search_results, playlist_id, playlist_name)
+
+@login_required
+def add_search_result_to_room(request, search_results, playlist_id, playlist_name):
+    if len(search_results['track_id']) == 1:
+        # Scenario 1: Add a new track using a valid Spotify track name - Happy Path
+        TRACK_ID = search_results['track_id']
+        add_track_id_to_playlist(request, playlist_id, playlist_name, TRACK_ID) 
+    elif len(search_results['track_id']) > 1:
+        # Scenario 2a: Search for a new track using an ambiguous Spotify track name - Happy Path
+        # Scenario 2b: Search for a new track to add to the playlist - Happy Path
+        context = search_results
+        return render(request, 'room.html', context)  
+    else:
+        messages.info(request, 'INFO: No tracks available.')
+        return redirect('room.html')
+    # Scenario 3: Add a new track using an invalid Spotify track name - Unhappy Path
+    # Handled by add_track_id_to_playlist
+    # Scenario 4: Add a duplicate track using a Spotify ID - Unhappy Path
+    # Handled by add_track_id_to_playlist
+
+@login_required
+def add_track_id_to_playlist(request, auth_manager, playlist_id, playlist_name, track_id):
+    cache_handler = spotipy.cache_handler.CacheFileHandler(cache_path=session_cache_path(request))
+    auth_manager = spotipy.oauth2.SpotifyOAuth(cache_handler=cache_handler)
+    spotify = spotipy.Spotify(auth_manager=auth_manager)
+
+    room = get_object_or_404(Room, pk=request.user.active_room_id, created_by=request.user, status=Room.ACTIVE, members__in=[request.user])
+
+    # TO DO: Tracks are stored in database at user_playlist_tracks function
+    if Track.objects.filter(track_id=track_id).count() == 0:
+        try: 
+            # Scenario 1: Add a new track using a valid Spotify ID - Happy Path
+            # {
+            #         "snapshot_id": "MzgsNjM1NDNkN2I0MGNlNzBjNDc0NGYxODAyOWVhZmEyNjFlZWQxYTZiZg=="
+            # }
+            snapshot_id = spotify.playlist_add_items(playlist_id=playlist_id, items=track_id, position=None)
+        except:
+            # Scenario 2: Add a new track using an invalid Spotify ID - Unhappy Path
+            # {
+            #     "error": {
+            #     "status": 400,
+            #     "message": "Payload contains a non-existing ID"
+            # }
+            messages.error(request, 'ERROR ' + request['error']['status'] + ': ' + request['error']['message'])
+            return redirect('room.html')
+        else:
+            messages.success(request, 'SUCCESS? Add track id to playlist attempted.')
+    else:
+        # Scenario 3: Add a duplicate track using a Spotify ID - Unhappy Path
+        messages.error(request, 'ERROR: Track already on playlist')
+        return redirect('room.html') 
+
+    user_playlist_tracks(request, playlist_id, playlist_name)
+
